+++ persistence/src/main/java/by/dmitrui98/dao/MaterialDao.java
@@ -5,6 +5,6 @@ import by.dmitrui98.entity.Material;
 /**
  * Created by Администратор on 29.04.2017.
  */
-public interface MaterialDao extends BaseDao<Material, Integer> {
+public interface MaterialDao extends BaseDao<Material, Long> {
     Material getByName(String name);
 }
